[ffx][pbms] fix --force removal
ffx product get tries to recognize the pb directory before removing it,
just to be safe. The dir wasn't being recognized because the PB included
more than the product bundle. This Cl changes the downloaded result to
be just a PB.

Also:fix error output if local output dir is unusable

When using ffx product get with --force allows for reusing the local
output directory. If this fails, a helpful error message should be
shown, but using .context() breaks the ffx_bail!() output.

A better fix would be to make .context() and ffx_bail!() cooperate,
but that's out of scope for my current work, so I'm removing some
.context() calls.

Change-Id: I93ed7547b55393c1f9ba236161446a341a3f2aa4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785773
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Dave Schuyler <dschuyler@google.com>