[nvme] Fix driver shutdown deadlock and re-enable tests.
In Namespace::IoLoop(), the driver shutdown boolean is checked in
between waiting for a sync completion and resetting it. Hence, it's
possible for both the driver shutdown boolean to be set and the
sync completion to be signalled (in Namespace::DdkRelease()) in the
narrow time window between after checking the driver shutdown
boolean but before resetting the sync completion, which results in
a deadlock.

The cancellation of pending IOs at driver shutdown time are handled
in Namespace::DdkRelease().

Bug: 118907
Multiply: fuchsia-pkg://fuchsia.com/nvme-tests#meta/nvme-unit-tests.cm
Change-Id: Ic47a8cd1a4e70df512b81153c5879c5ac842ec4d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787307
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: HanBin Yoon <hanbinyoon@google.com>