[emu] Enable custom virtual devices for emu
This CL adds logic to the handling of the "--device" flag to allow the
user to specify a custom virtual device file path. If the input to
"--device" is not a path to a file, `ffx emu` will continue to treat it
as the name of a virtual device embedded in the target product bundle.

Bug: 108747
Change-Id: Ieb1424171ef8be1c3964a614e31da10dafb4e863
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778324
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>