[ffx component explore] Convert built-ins to trampolines
Now that trampolines are enabled for --tools, this CL converts all of
the built-in binaries enabled for the Dash shell to use trampolines.

Also, this CL removes several integration tests because the test
harness doesn't provide the limited fuchsia.process.Resolver and
fuchsia.kernel.VmexResource capabilities. With the conversion of the
built-in shell binaries to trampolines, all binaries now need these
capabilities to run.

Instead, new systems test will be added in fxr/779104.

Bug: 116646
Change-Id: If04ec202cc0d6ee8d958e9130c40adc2b0b3d8a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778022
Commit-Queue: Gary Boone <gboone@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>