[starnix] Extend MAP_GROWSDOWN mapping in response to user faults
This implements a portion of the MAP_GROWSDOWN feature that extends a
mapping to cover a faulting access from userspace. To support this the
memory manager has a feature to extend mappings for a single address
which is invoked when receiving a fault for a not-present page.

This should eventually be the only mechanism for growing such regions,
however since Starnix<->restricted copies currently happen through VMO
operations and not mapped memory they will need a separate codepath that
performs growth over a range of addresses to work properly. This change
also does not implement guard regions or the PROT_GROWSDOWN flag.

Bug: 97514

Change-Id: I1f8f0ac2b6a2d277ab73605b07b2a0b30665e694
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786430
Commit-Queue: James Robinson <jamesr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>