[account_handler] Additional success logging
Its been a little tricky to debug the account integration since log
messages were not consistently present or structured. This just goes
through all the FIDL method handlers with the goal of making sure we
got an informative message with the name of the method for every code
path.

Change-Id: I6e33f29cb26a0d10935c6cfe392d4382c99d0f53
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787049
Reviewed-by: Angad Pal Singh <apsbhatia@google.com>
Reviewed-by: James Buckland <jbuckland@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>