[fbl] assert on BitField overflow
This commits promotes a ZX_DEBUG_ASSERT to a ZX_ASSERT in
BitFieldElement's operator=, disallowing setting values that are greater
than the maximum.

BitFieldElement operator= allows assigning a value to a BitField, and
there was a DEBUG_ASSERT that said value is less than the maximum possible
value for that element. However, in builds without debug asserts, the
code would still set the value, affecting bits of the bitfield unrelated
to the current element.

Test: fx test "fuchsia-pkg://fuchsia.com/fbl-test#meta/fbl-test.cm"
Bug: 115898
Change-Id: I4e972b91f6756c7e4fa3096c749bac74f92d15fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764255
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>