[chromium] Make canonical target empty
A canonical target is required per
https://fuchsia.dev/fuchsia-src/development/source_code/layout?hl=en#canonical_targets,
but nothing should depend on it for specific package(s). An empty
target addresses the requirement while not affecting product behavior
or needing to decide what it should include.

This removes one TODO for closed fxbug.dev/81925. Also update the
tracking issue for the other such TODO.

Move the canonical target to the BUILD.gn file for the directory where
it is needed.

Bug: 119075, 81925, 115624
Change-Id: I4c3e7c1d44695d7e88497256d97c59589d0de423
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786551
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: Chong Gu <chonggu@google.com>
Reviewed-by: James Robinson <jamesr@google.com>