[starnix] Handle closed handles in epoll signaling.
Handle the case where a file is signaled but closed out from under
epoll. In this case the wait should continue/timeout rather than
returning an error. Add tests for this case. (Note the general issue of
closed files in epoll in issue 64296 is still partially unresolved, this
change only handles the case where they're in the triggered list).

Internally pass the max events to epoll_wait() as a usize to make clear
that it's been checked to be nonnegative. Add tests for this case.

Add tests for invalid size passed to epoll_create().

Bug: 64296, 117979

Change-Id: I1c9dd50ff5081b662fbae718af977e31ce7c0da1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780503
Reviewed-by: Benjamin Lerman <qsr@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>