[driver_manager] Another `has_outgoing_directory` -> `uses_fragment_driver`
Since we are specifically looking for a fragment that uses the fragment
driver, this is the appropriate check rather than whether the bound
device has an outgoing directory. These conditions used to be equivalent
but will not be in the future when we enable dual Banjo/FIDL nodes.

Bug: 116754
Change-Id: Id38c3971b087e7e7185a8ed40751f63ef98411d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787902
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Fuchsia-Auto-Submit: Ian Fisher <iafisher@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>