[diagnostics][hierarchy] Bring back payload:null and align a few cases
Turns out we had an optimization for feedback snapshots in which we
don't include hierarchies with `payload: null`. Therefore I think we
should prefer that over `payload: { root: {} }` when we fully filter a
hierarchy.

This change brings alignment for that situation.

The tests reflect the cases that I consider to not be aligned before.

Run-All-Tests: true

Change-Id: I928cd75ecf29319887e17834bce12289d7f4995e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788324
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>