[network/tests] Fix Flake in DAD integration tests
This test flaked because the test waited more than one second between
adding the address, and sending a conflicting response to the neighbor
solicitation. As a result, DAD (unexpectedly) finished successfully.

Unfortunately this category of flake would be very difficult to *solve*,
since we'd need to build the Netstack with a fake timer that could be
manipulated by the test. I opted to instead move the execution of
`dad_fn` sooner after the address is added, to significantly decrease
the likely hood of DAD finishing before the test injects its response.

Fixed: 119164

Multiply: netstack-ipv6-integration-test
Change-Id: I7bea7b60fe0590dc9c9aa4866c9fd3cd1cdda36f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787805
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>