[netstack3] Serialize TCP segments lazily
Add a method to BufferProvider that allocates without reuse and use it
to implement Serializer for SendPayload. The new method is necessary
because SendPayload is only a view and doesn't hold a buffer that can be
reused. Implementing Serializer for SendPayload allows it to be passed
in as the inner body of a packet.

Previously, the contents of the SendPayload were being copied into a
Vec, which was then being encapsulated in other packet builders. This
resulted in a double allocation: once for SendPayload -> Vec, then again
to create a new Vec to hold the entire frame/packet. This change removes
the first allocation by deferring copying from the SendPayload until the
final Vec is allocated for the entire serializer.

Change-Id: I9d8ae75af1cb6004e0cf855d7e6c68debd5ab994
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786703
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>