[network] Add `event_stream_from_state` helpers
This method instantiates either a `WatcherV4` or a `WatcherV6` client,
and returns events from the watcher as a `Stream`. This abstracts over
the Hanging-Get style API.

Additionally this CL introduces fake implementations of the `StateV4`,
`StateV6`, `WatcherV4` and `WatcherV6` protocols as part of a `testutil`
mod. This was necessary to test `event_stream_from_state` and may prove
useful to future clients.

Bug: 115870
Change-Id: I2b83edc7e2dbbba94f3a28416b5b41078c0e139f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762867
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>