[build][bazel] Make the Bazel TOPDIR more configurable
The location of the Bazel top directory, relative to the Ninja output
dir is now officially determined by the value store in a configuration
file at `//build/bazel/config/main_workspace_top_dir`, which is read
by GN and shell or python scripts alike.

This determines the location of the workspace/ and output_base/
directories for the Bazel workspace, and a few other things.

All other references to this sub-dir are removed (except in comments
and documentation).

This will allow moving the workspace to a new location in the future,
with a very simple CL that changes the value of that file (and
comments + documentation).

+ Add tools/devshell/lib/bazel_utils.sh to export the fx-bazel-top-dir
  function, then make 'fx bazel` and `fx get-workspace-dir` use it.

Bug: 113116

Change-Id: If2914120a824118dd243882e77124c7098b28924
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779045
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Shai Barack <shayba@google.com>