[emu] Tweak path handling logic for --device
The attached bug describes how running `ffx emu start` with no --device
flag prints a message about using a path as a virtual device file, which
makes no sense to a user who didn't enter such a path.

This cl adjusts the handling of the Some("") case for the --device flag,
such that it is treated the same as None. It also short-circuits the
parsing logic when the path given is not a valid file.

Bug: 119271
Change-Id: Ie123dd95ea912afc139cf246e2ec87aa5deea809
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788422
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>