[stress-tests][fxfs] Add progress timer
If no files are created after 60s, a timer will trip that will cause the
test to fail.  This helps us ensure that forward progress is made (which
should be guaranteed since we are also deleting files in another task).

60s is arbitrary; it's big enough that flakes are unlikely.  We could go
even higher if we wanted to, since these tests run for fairly long
periods of time.  The period at which the `crash` variant restarts the
test environment is increased to 90s so that the timer isn't completely
useless for that test.

Fixed: 110593
Change-Id: Iaf93d6329a0f0fc7eb376015c80cb6efe96b18a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/783882
Reviewed-by: Stephen Demos <sdemos@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>