[scenic][scheduling] Add a DCHECK to catch ordering issues
Adds a DCHECK to DefaultFrameScheduler checking that RegisterPresent()
happens in PresentId order. We depend on this ordering throughout
DefaultFrameScheduler and when it broke root causing the issue was
very difficult. This should help with that in the future.

Fixed: 119268
Change-Id: I879548e1e85888bb8cf97413124ca99a43b92a9e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789074
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jaeheon Yi <jaeheon@google.com>
Fuchsia-Auto-Submit: Mikael Pessa <mikaelpessa@google.com>