[pager][fxfs] Support zero range bit on writeback
Add support to the rust pager wrappers for passing a client's
expectation that a range is entirely zero when the client begins to
clean a range.

This is used to avoid a race where a page can be dirtied between
query_dirty_ranges returning that the page is all zeroes and calling
writeback_begin with the page.

Change-Id: I8065a3b2d4e44e16b80627383efefaa64cff5420
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785907
Reviewed-by: Chris Tam <godtamit@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>