[component][fuzzing] Check for message limits in ChildProcess
This CL adds checks to `ChildProcess` methods such as `AddArg`,
`SetEnvVar`, and `AddChannel` to ensure the parameters passed to
`fdio_spawn` can fit in a FIDL message.

Previously, the easiest way to exceed these limits and thereby fail to
spawn a process was to provide a libFuzzer-based binary with a large
number of individual inputs to try. When run as a unit test, the engine
tries all inputs in the seed corpus. Simply providing a sufficiently
large seed corpus would cause tests to fail.

As a result, the LibFuzzerRunner has also been updated to use the
feedback from its ChildProcess object to batch inputs into several tries
when they would not fit into one.

As a result of testing, a bug was observed and corrected where the
fuzzer tests were not converting relative paths to be package-absolute.

Finally, in order to adequately test all of this, the sample corpus
under common/testing/data was dramatically increased to include 2000
separate inputs.

Change-Id: Id934f93b84d8a029815f98c19e18feb7abab054b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788428
Reviewed-by: Cameron Finucane <eep@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>