[starnix] Handle ptmx not being in /dev/pts
Depending on the distribution, the ptmx file can either be a symlink to
/dev/pts/ptmx, or another device node. Because it must access /dev/pts,
instead of relying on the node being on the right file system, this CL
pass the root node from the /dev/pts filesystem to the node operations
on ptmx. This is tested by BasicPtyTest.ReplicaEntryGoneAfterMasterClose
when ptmx is a device on /dev instead of a symbolic link.

Change-Id: Icad1581395cf23bf49c6ebb414e3017ec205ecd5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788336
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>