[audio][drivers][as370] Add 96kHz support
Adds support for 96kHz and switch the advertised format
from 48kHz to 96kHz. We adjust the pre-multiplier shift
to match the mic gain in other products.

For output we change the divisor in the clock tree
changing BCLK from 3.072MHz to 6.144MHz. Note that this
changes the amount per DMA from 21.3msecs to 10.67msecs.

For input the dividers stay the same to keep the DMICs
clock at 3.072MHz but we change the PDM input bits per
sample from 64 to 32, and also reduce the dma_mtus to
use switch from 21.3msecs to 10.67msecs DMAs.

See also go/tq-as370-pdm-pcm-perf.

Bug: 107645
Test: as370-tdm-output-tests
Test: as370-pdm-input-tests
Test: audio-dsp-test
Test: audio-driver-ctl -r 96000 -b 24 tone in Pinecrerst
Test: audio-driver-ctl -r 96000 record /tmp/t in Pinecrest

Change-Id: I47bd6d1de42a457df88f9cc82b0be21ab1b87e46
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787426
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>