[emu] Swap Mutex guards for serial tests
The attached bug explains how we recently discovered that using automock
on async functions is unstable, and attempting to work around this with
a Mutex is insufficient since the mutex is thread-owned and thus does
not guard the static mock contexts between the various tests.

This CL removes the mutex and replaces it with annotations to serial the
tests that utilize the mocking context instead. By serializing the
tests, we give them atomicity and ensure that the context is set, used,
and destroyed all before the next test starts. While this can slow down
the test execution, these are very fast tests and the delay is not
enough to notice or care.

Bug: 119219
Change-Id: Ib53464a225ac7e4d43663ac0fa54ff001fa5796d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789178
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>