[netstack3] Put Keep-Alive and Nagle options into a single struct
This gives us the benefit of not adding more and more accessors for each
of the socket options that we add in the future. This change also
removes the `NagleAlgorithm` enum and uses a boolean as a named field.
It does not worsen readability and aligns with existing mechanism used in
enabling keep-alive.

Bug: 117115
Change-Id: I4cff6f2406bf61faeed12450d5ff74600a0c202d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789262
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>