[tefmocheck] Suspend test may cause an OOPS while passing.
The suspend test checks that device successfully reaches the S3 (Suspend
To Ram) power state and then attempts to resume. At this stage we do not
expect resume to succeed and in fact it is likely there will be multiple
OOPSes caused by leaking BTIs. This does not reflect a failure of the
test, but simply the expected state of resume currently.

Bug: 117528
Change-Id: I72dc8857b72c054bb3b6f37da90a4e83b59bf81e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788273
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Sarah Pham <smpham@google.com>