[compat] Refactor to ExportAfterInit
Hopefully simplify the device lifetime a bit.

Export is renamed to ExportAfterInit and is only called when the
device's init has finished.

InitReply now waits for the parent's init to finish, so we always know
that our parent has been initialized and exported.

This means that we should always be in a state where ExportAfterInit
can directly export to devfs and create the node.

This means that NON_BINDABLE devices should have the same flow as
bindable devices.

This also means that DevfsExport::MakeVisible is no longer needed.

This fixes a bug in the ramdevice library where the code does not wait
for the initial ramdevice to exist in devfs. This bug was exposed
because this CL changes the timing between the parent creating the
device and it appearing in devfs.

Multiply: compat_test
Bug: 117760
Change-Id: Ia3e167cbb80d44e40361639c878f8e82b36362a7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789004
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Ian Fisher <iafisher@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>