[archivist] Gracefully handle invalid UTF8 in log messages
Instead of returning an error when getting invalid UTF8, we can instead
replace the invalid UTF8 sequences with �.

The Archivist will no longer close log sockets when it fails to parse a
message received through the socket, instead it will leverage the
existing logic handling errors in StoredMessage.parse to return a Data
with an error set explaining that the message couldn't be parsed. When
invalid UTF-8 is found, no error will be set, but the message will
contain the � for each of the invalid UTF-8 sequences.

Fixed: 119194

Change-Id: I1a7688d73ad498bddceb4516a69801afe3b06bd4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790122
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>