[input] Don't block input pipeline if reader stuck
When an event is handled and sent to registered listeners, it is possible a listener's reader might get stuck and never send a response. Because of this, we want `handle_input_event` to return after sending the event and we want to make sure other listeners can still receive the event if a reader gets stuck and that subsequent events can still be handled.

Fixed: 106843
Change-Id: Ic806410da64bd008995964e775e2c9b9349208f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771343
Fuchsia-Auto-Submit: Justin Kobler <jkobler@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Caroline Liu <carolineliu@google.com>