[debugger] Improve the UI while loading lots of symbols
When loading and indexing very large symbol files (chromium, for
example), the user is just stuck with a non-responsive [zxdb] prompt,
which is not ideal.

This change aims to make that a little nicer by showing a "Loading
symbols for <process name>..." prompt that continuously shows the user
that *anything* is happening. Since there's really not a ton that can be
done before symbols are loaded, it's better to just not even present a
prompt than to move the indexing work to a background thread and show
them a responsive console.

FIXED: 6311
Change-Id: Ib5356ff98cbb0c2d6520f5ae855145169e08fb3f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/783302
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>