[bt-hci-marvell] Initialize mac address
On startup, read the mac address from boot data and pass it to the
controller. The driver will wait for confirmation from the
controller before sending the ddkInit status response.

To support this, we use a driver-specific vendor command zx:channel
that is distinct from the other zx::channel's we use for host
communication. This will allow us to use the standard event loop
for processing driver-generated commands.

Test: fx test bt-hci-marvell-tests
Change-Id: I30caadee01fca0703c1b6f466852c67e78c303aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779804
Reviewed-by: Ben Lawson <benlawson@google.com>
Commit-Queue: Joshua Conner <joshconner@google.com>