[reachability] Adding integration tests using FIDL
Create a reachability component that incorporates the fake clock library
so that the test environment can be accelerated. The named_timeout
allows the code to pause prior to hitting the timeout window. Prior to
timing out, the code is accelerated, then slowed down to real time.

The integration tests cases where the internet_available flag was
initially true or initially false. This helps simulate the cases where
the internet went from good to bad or bad to good.

Additionally, testing of edge cases, such as the method of interacting
with the API, are included.

Change-Id: Ie0a47c620296e53356ad1e4688c17ff58a27febd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779254
Reviewed-by: Dylan Swiggett <swiggett@google.com>
Fuchsia-Auto-Submit: Antonio Linhart <antoniolinhart@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>