[starnix] Pass galaxy outgoing dir to kernel
This change updates starnix_manager to pass the outgoing directory for
dynamic galaxies to the starnix_kernel.

This will allow the kernel to serve a component runner on behalf of the
galaxy.

This also removes the code passing the component controller from
starnix_manager to the kernel, since a future change will update the
manager to serve the component controller.

Change-Id: I21c041ff9fc9da5186cf03c37535c073fb652fb6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790922
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>